김민형 / 회사 수정 완

+++ src/main/java/com/my/attendance/domain/Company.java
@@ -7,7 +7,7 @@ public class Company {
 	private String companyId;
 	private String comPw;
 	private String comName;
-	private int comRegno;
+	private long comRegno;
 	private String comAddr;
 	private String comDetailAddr;
 	private int comPostcode;
